# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Community Growth Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/bulk:
    post:
      operationId: bulkCreateContacts
      tags: [contacts]
      summary: Bulk create contacts from text or CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkContactInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/bulk-delete:
    post:
      operationId: bulkDeleteContacts
      tags: [contacts]
      summary: Bulk delete contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Groups
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    put:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /groups/{id}/add-contacts:
    post:
      operationId: addContactsToGroup
      tags: [groups]
      summary: Add contacts to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddContactsInput"
      responses:
        "200":
          description: Addition started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Campaigns
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - createdAt
    ContactInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
          nullable: true
        notes:
          type: ["string", "null"]
      required:
        - name
        - phone
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
          nullable: true
        notes:
          type: ["string", "null"]
    BulkContactInput:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    BulkImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - skipped
        - errors
    BulkDeleteInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
    Group:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: string
        name:
          type: string
        memberCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - name
        - createdAt
    GroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        groupId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        memberCount:
          type: ["integer", "null"]
      required:
        - name
    AddContactsInput:
      type: object
      properties:
        contactIds:
          type: array
          items:
            type: integer
      required:
        - contactIds
...