# lib/api-zod/src/generated/types/bulkContactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Community Growth Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkContactInput {
  text: string;
}