# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Community Growth Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  tags?: string[] | null;
  /** @nullable */
  notes?: string | null;
}